sdl3_image_sys
Module image

Header file for SDL_image library

A simple library to load images of various formats as SDL surfaces

Structs

IMG_Animation

Animated image support

Constants

SDL_IMAGE_MAJOR_VERSION

Printable format: “%d.%d.%d”, MAJOR, MINOR, MICRO

SDL_IMAGE_MICRO_VERSION

SDL_IMAGE_MINOR_VERSION

SDL_IMAGE_VERSION

This is the version number macro for the current SDL_image version.

Functions

IMG_FreeAnimation⚠

Dispose of an IMG_Animation and free its resources.

IMG_Load⚠

Load an image from a filesystem path into a software surface.

IMG_LoadAVIF_IO⚠

Load a AVIF image directly.

IMG_LoadAnimation⚠

Load an animation from a file.

IMG_LoadAnimationTyped_IO⚠

Load an animation from an SDL datasource

IMG_LoadAnimation_IO⚠

Load an animation from an SDL_IOStream.

IMG_LoadBMP_IO⚠

Load a BMP image directly.

IMG_LoadCUR_IO⚠

Load a CUR image directly.

IMG_LoadGIFAnimation_IO⚠

Load a GIF animation directly.

IMG_LoadGIF_IO⚠

Load a GIF image directly.

IMG_LoadICO_IO⚠

Load a ICO image directly.

IMG_LoadJPG_IO⚠

Load a JPG image directly.

IMG_LoadJXL_IO⚠

Load a JXL image directly.

IMG_LoadLBM_IO⚠

Load a LBM image directly.

IMG_LoadPCX_IO⚠

Load a PCX image directly.

IMG_LoadPNG_IO⚠

Load a PNG image directly.

IMG_LoadPNM_IO⚠

Load a PNM image directly.

IMG_LoadQOI_IO⚠

Load a QOI image directly.

IMG_LoadSVG_IO⚠

Load a SVG image directly.

IMG_LoadSizedSVG_IO⚠

Load an SVG image, scaled to a specific size.

IMG_LoadTGA_IO⚠

Load a TGA image directly.

IMG_LoadTIF_IO⚠

Load a TIFF image directly.

IMG_LoadTexture⚠

Load an image from a filesystem path into a GPU texture.

IMG_LoadTextureTyped_IO⚠

Load an image from an SDL data source into a GPU texture.

IMG_LoadTexture_IO⚠

Load an image from an SDL data source into a GPU texture.

IMG_LoadTyped_IO⚠

Load an image from an SDL data source into a software surface.

IMG_LoadWEBPAnimation_IO⚠

Load a WEBP animation directly.

IMG_LoadWEBP_IO⚠

Load a WEBP image directly.

IMG_LoadXCF_IO⚠

Load a XCF image directly.

IMG_LoadXPM_IO⚠

Load a XPM image directly.

IMG_LoadXV_IO⚠

Load a XV image directly.

IMG_Load_IO⚠

Load an image from an SDL data source into a software surface.

IMG_ReadXPMFromArray⚠

Load an XPM image from a memory array.

IMG_ReadXPMFromArrayToRGB888⚠

Load an XPM image from a memory array.

IMG_SaveAVIF⚠

Save an SDL_Surface into a AVIF image file.

IMG_SaveAVIF_IO⚠

Save an SDL_Surface into AVIF image data, via an SDL_IOStream.

IMG_SaveJPG⚠

Save an SDL_Surface into a JPEG image file.

IMG_SaveJPG_IO⚠

Save an SDL_Surface into JPEG image data, via an SDL_IOStream.

IMG_SavePNG⚠

Save an SDL_Surface into a PNG image file.

IMG_SavePNG_IO⚠

Save an SDL_Surface into PNG image data, via an SDL_IOStream.

IMG_Version⚠

This function gets the version of the dynamically linked SDL_image library.

IMG_isAVIF⚠

Detect AVIF image data on a readable/seekable SDL_IOStream.

IMG_isBMP⚠

Detect BMP image data on a readable/seekable SDL_IOStream.

IMG_isCUR⚠

Detect CUR image data on a readable/seekable SDL_IOStream.

IMG_isGIF⚠

Detect GIF image data on a readable/seekable SDL_IOStream.

IMG_isICO⚠

Detect ICO image data on a readable/seekable SDL_IOStream.

IMG_isJPG⚠

Detect JPG image data on a readable/seekable SDL_IOStream.

IMG_isJXL⚠

Detect JXL image data on a readable/seekable SDL_IOStream.

IMG_isLBM⚠

Detect LBM image data on a readable/seekable SDL_IOStream.

IMG_isPCX⚠

Detect PCX image data on a readable/seekable SDL_IOStream.

IMG_isPNG⚠

Detect PNG image data on a readable/seekable SDL_IOStream.

IMG_isPNM⚠

Detect PNM image data on a readable/seekable SDL_IOStream.

IMG_isQOI⚠

Detect QOI image data on a readable/seekable SDL_IOStream.

IMG_isSVG⚠

Detect SVG image data on a readable/seekable SDL_IOStream.

IMG_isTIF⚠

Detect TIFF image data on a readable/seekable SDL_IOStream.

IMG_isWEBP⚠

Detect WEBP image data on a readable/seekable SDL_IOStream.

IMG_isXCF⚠

Detect XCF image data on a readable/seekable SDL_IOStream.

IMG_isXPM⚠

Detect XPM image data on a readable/seekable SDL_IOStream.

IMG_isXV⚠

Detect XV image data on a readable/seekable SDL_IOStream.

SDL_IMAGE_VERSION_ATLEAST

This macro will evaluate to true if compiled with SDL_image at least X.Y.Z.