sdl3_image_sys::image
Function IMG_isAVIF

pub unsafe extern "C" fn IMG_isAVIF(src: *mut SDL_IOStream) -> bool

Detect AVIF image data on a readable/seekable SDL_IOStream.

This function attempts to determine if a file is a given filetype, reading the least amount possible from the SDL_IOStream (usually a few bytes).

There is no distinction made between “not the filetype in question” and basic i/o errors.

This function will always attempt to seek src back to where it started when this function was called, but it will not report any errors in doing so, but assuming seeking works, this means you can immediately use this with a different IMG_isTYPE function, or load the image without further seeking.

You do not need to call this function to load data; SDL_image can work to determine file type in many cases in its standard load functions.

Parameters

src: a seekable/readable SDL_IOStream to provide image data.

Return value

Returns non-zero if this is AVIF data, zero otherwise.

Availability

This function is available since SDL_image 3.0.0.

See also

IMG_isAVIF

IMG_isICO

IMG_isCUR

IMG_isBMP

IMG_isGIF

IMG_isJPG

IMG_isJXL

IMG_isLBM

IMG_isPCX

IMG_isPNG

IMG_isPNM

IMG_isSVG

IMG_isQOI

IMG_isTIF

IMG_isXCF

IMG_isXPM

IMG_isXV

IMG_isWEBP